% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyBib-package.R
\docType{package}
\name{LazyBib-package}
\alias{LazyBib}
\alias{LazyBib-package}
\title{LazyBib: Bib File Wrangling}
\description{
Bib file wrangling.
}
\author{
\strong{Maintainer}: Yangzhuoran Fin Yang \email{yangyangzhuoran@gmail.com} (\href{https://orcid.org/0000-0002-1232-8017}{ORCID})

}
\keyword{internal}
